feat: test the use of offline messages stored in mysql
